+++ src/pages/gateway/release-notes.md
@@ -101,7 +101,7 @@ This release contains the following changes to API Mesh:
 ### Enhancements
 
 - [GraphQL aliasing](https://graphql.org/learn/queries/) is now enabled.
-  - As a part of this enhancement, several [handler](../reference/handlers/index.md) and [transform](../reference/transforms/index.md) packages have been updated.
+  - As a part of this enhancement, several [handler](././handlers/index.md) and [transform](./transforms/index.md) packages have been updated.
 
 ## August 30, 2023
 
@@ -128,7 +128,7 @@ This release contains the following changes to API Mesh:
 
 ### Enhancements
 
-- Added the [SOAP handler](../reference/handlers/soap.md) to API Mesh.
+- Added the [SOAP handler](./handlers/soap.md) to API Mesh.
   - The SOAP handler is experimental and should not be used in production deployments.
 - Added internal logging improvements.